+++ ui/src/app/components/buttons.tsx
@@ -1,7 +1,7 @@
 import {styles as cardStyles} from 'app/components/card';
 import {ClrIcon} from 'app/components/icons';
 import {TooltipTrigger} from 'app/components/popups';
-import {IconComponent} from 'app/icons/icon/component';
+import {IconComponent} from 'app/icons/icon';
 import colors from 'app/styles/colors';
 import * as fp from 'lodash/fp';
 import * as React from 'react';
