+++ ui/src/app/icons/icons.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from '@angular/core';
 
-import {EditComponent} from 'app/icons/edit/component';
+import {EditComponent} from 'app/icons/edit';
 import {ExpandComponent} from 'app/icons/expand/component';
 import {NotebookComponent} from 'app/icons/notebook/component';
-import {ReminderComponent} from 'app/icons/reminder/component';
+import {ReminderComponent} from 'app/icons/reminder';
 import {ShareComponent} from 'app/icons/share/component';
 import {ShrinkComponent} from 'app/icons/shrink/component';
 import {TrashComponent} from 'app/icons/trash/component';
